+++ packages/agtree/CHANGELOG.md
@@ -8,6 +8,18 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html
 
+## [2.1.4] - 2024-11-25
+
+### Added
+
+- Error messages for rules parsing errors.
+
+### Fixed
+
+- Missing children data in the deserializer for certain nodes.
+
+[2.1.4]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-v2.1.4
+
 ## [2.1.3] - 2024-10-21
 
 ### Fixed